requirements.txt
# requirements.txt

-i https://pypi.org/simple
asgiref==3.8.1 ; python_version >= '3.8'
django==4.2.13
psycopg[binary]==3.1.19
psycopg-binary==3.1.19
sqlparse==0.5.0 ; python_version >= '3.8'
typing-extensions==4.12.2 ; python_version >= '3.8'
whitenoise==6.1.0
gunicorn==20.1.0
pytest==8.0.0
pytest-django==4.5.2
environs[django]==9.5.0
openai==1.35.7
django-cleanup==8.1.0
pillow==10.4.0
django-crispy-forms==2.2
crispy-bootstrap5==2024.2
celery==5.4.0
django-celery-results==2.5.1
redis==4.5.1  
django-imagekit==5.0.0
requests==2.32.3